WIP refacto general spell meta
TODO refacto Conversion

optionnel : refaire tout main activity en modul et spellCastActiv aussi
et faire plus de creation dans des layout que dans le code (cf fragment2 par spell)
