+++ app/src/main/java/stellarnear/yfa_dealer/Spells/_oldMeta_Check_Img.java
@@ -8,11 +8,11 @@ import android.widget.ImageButton;
  * Created by jchatron on 29/11/2017.
  */
 
-public class Meta_Check_Img extends AppCompatActivity {
+public class _oldMeta_Check_Img extends AppCompatActivity {
     CheckBox checkbox;
     ImageButton ImgBu;
 
-    public Meta_Check_Img( CheckBox checkbox,ImageButton ImgBu)
+    public _oldMeta_Check_Img( CheckBox checkbox,ImageButton ImgBu)
     {
         this.checkbox = checkbox;
         this.ImgBu = ImgBu;
