+++ app/src/main/java/stellarnear/yfa_dealer/Spells/_oldPair_Meta_Rank.java
@@ -4,18 +4,18 @@ package stellarnear.yfa_dealer.Spells;
  * Created by jchatron on 29/11/2017.
  */
 
-public class Pair_Meta_Rank {
+public class _oldPair_Meta_Rank {
     int rank;
-    Meta_Check_Img meta;
+    _oldMeta_Check_Img meta;
 
-    public Pair_Meta_Rank( Meta_Check_Img meta,int rank)
+    public _oldPair_Meta_Rank(_oldMeta_Check_Img meta, int rank)
         {
             this.meta = meta;
             this.rank = rank;
         }
 
 
-       public Meta_Check_Img getMeta(){
+       public _oldMeta_Check_Img getMeta(){
             return this.meta;
         }
